patch qkv_rot
